+++ src/pylhe/__init__.py
@@ -515,6 +515,10 @@ class LHEInitRWGT:
             else:
                 yield from entry.weights
 
+    def list_weights_ids(self) -> list[str]:
+        """Return all weight IDs in declaration order."""
+        return [w.id for w in self.iter_weights()]
+
     def weights_by_id(self) -> dict[str, LHEInitRWGTWeight]:
         """Return a dictionary mapping weight IDs to LHEInitRWGTWeight instances for all weights in the <initrwgt> block."""
         return {w.id: w for w in self.iter_weights()}
@@ -1142,6 +1146,8 @@ class LesHouchesEvents:
 
         if isinstance(fileobject, h5py.File):
             init = lheh5.read_init(fileobject)
+            header = lheh5.read_header(fileobject)
+            comment = lheh5.read_comment(fileobject)
 
             def _hdf5_generator() -> Iterator[LHEEvent]:
                 with fileobject as h5:
@@ -1151,6 +1157,8 @@ class LesHouchesEvents:
             return LesHouchesEvents(
                 init=init,
                 events=events if generator else list(events),
+                header=header,
+                comment=comment,
                 version=None,  # We leave the version as None since HDF5 versioning is unrelated to LHE XML versioning.
             )