+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- LHEH5 files now store generator and root-comment metadata under `/metadata`.
+
 ## [2.0.0] - 2026-07-13
 
 ### Added